Module Caqti_driver_mariadb

Caqti driver for MariaDB (bindings).

This driver is implemented in terms of the mariadb OPAM package. It handles URIs of the form

mariadb://<user>:<password>@<host>:<port>/?socket=<socket>

where components are optional and present ones are passed on to the correspondingly named arguments to the connect function of MariaDB.

The interface provided by this module should normally not be used by applications, but provides access to some MariaDB specifics in case they are needed.

Error Details

The following provides access to diagnostics collected from the MariaDB connection and statement objects.

type Caqti_error.msg +=

| Error_msg of {

errno : int;	(*
The error number returned by mysql_errno or mysql_stmt_errno.
*)

error : string;	(*
The error message returned by mysql_error or mysql_stmt_errno.
*)

}